Rename files
  ref #77
