+++ src/standardRule/build-validator.ts
@@ -1,5 +1,5 @@
 import { Validator } from "./validator"
-import { Utils } from "./standard-rule-utils"
+import { Utils } from "./utils"
 import { OrderType, Order } from "./order"
 import { Board, Unit } from "./types"
 import * as Error from "./error"
